Rename Python package to `rgain3`
