+++ rgain3/util.py
