+++ scripts/replaygain
@@ -1,7 +1,6 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
+#!/usr/bin/env python3
 
-from rgain.script.replaygain import replaygain
+from rgain3.script.replaygain import replaygain
 
 if __name__ == "__main__":
     replaygain()
