+++ setup.cfg
@@ -8,12 +8,12 @@ max-line-length = 80
 combine_as_imports = true
 default_section = THIRDPARTY
 include_trailing_comma = true
-known_first_party = rgain
+known_first_party = rgain3
 line_length = 80
 multi_line_output = 3
 not_skip = __init__.py
 
 [tool:pytest]
-addopts = --showlocals --full-trace --flake8 --isort --cov rgain --cov-report term-missing
+addopts = --showlocals --full-trace --flake8 --isort --cov rgain3 --cov-report term-missing
 flake8-max-line-length = 80
 flake8-max-complexity = 20
