+++ test/test_rgcalc.py
@@ -7,7 +7,7 @@ gi.require_version("Gst", "1.0")
 
 from gi.repository import GObject, Gst  # noqa isort:skip
 
-from rgain import GainType, rgcalc, util  # noqa isort:skip
+from rgain3 import GainType, rgcalc, util  # noqa isort:skip
 
 Gst.init([])
