+++ tox.ini
@@ -4,8 +4,7 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py27
-# py35 is very close to working
+envlist = py35,py36,py37
 
 [testenv]
 # necessary to make cov find the .coverage file
@@ -14,7 +13,7 @@ usedevelop = true
 # necessary to find pygobject which is impossible to install in a virtualenv :(
 sitepackages = true
 commands =
-    py.test -rf -l -s -x  --cov-report term-missing --cov rgain
+    py.test -rf -l -s -x  --cov-report term-missing --cov rgain3
     flake8 --max-complexity=10 rgain test
 deps =
     -rtest-requirements.txt
